support for M.[ foo ], M.{ foo }, etc.
This is a proposal implementation of the approach suggested by Jérémie
in github issue #10. I had to protect the expr production

    | m = a_UIDENT; "."; l = SELF -> <:ident< $uid:m$.$l$ >>

with a TRY around [ "."; SELF ], because otherwise it would fire
in the M.<delim> case, eat the last dot, and then fail with a parse
error on the delimiter.