Fix Makefile wrt SAVED_BINDIR and SAVED_LIBDIR
Export BINDIR and LIBDIR via env, so build/install.sh reads the
correct ones. The Makefile reads config.sh, which properly
double-quotes path strings, and GNU Make will then export variables
with double-quotes; that's fine, because build/install.sh itself reads
config.sh, which reads the variables the way they are intended to
be... except for BINDIR and LIBDIR that are kept from the environment if
defined, so they can be overridden by the Makefile. This change
ensures that they should be properly and not overly quoted.