fix usage of DEFINE-d identifiers in patterns in standard syntax
fixes #148

Note that both parameter-less macros (as in the bug report) and
parametrized macros are affected.

I wasn't able to find a testsuite in Camlp4, so I did not add
a test. I believe that the regression potential of this change is
fairly low: it only affects the action of DEFINE in patterns (which
are currently broken in standard syntax), not in expressions, and
I tested it in both standard and revised syntaxes.

Testcase for standard syntax:

    DEFINE FOO = ""
    DEFINE ID(x) = x

    let test x = match x with
    | Some(FOO) -> Some(FOO)
    | Some(ID "bar") -> Some(ID "bar")
    | Some _ -> Some "baz"
    | None -> None

Testcase for revised syntax:

    DEFINE FOO = "";
    DEFINE ID(x) = x;

    value test x = match x with
    [ Some FOO -> Some FOO
    | Some (ID "bar") -> Some (ID "bar")
    | Some _ -> Some "baz"
    | None -> None ];